Add editorconfig and modularize paradox-app example
